Update new module directory check
The check has been updated to be earlier in the process, directly after
getting the module's name. All the information required for it is
available at that point, and getting an error because of it later on is
annoying.

Additionally, the check has been updated. In case the path is already
taken, an error will be shown, and the user is given the option to give
another name.